Creating Data Processing GUI initial
This is the first commit of the Data Processing GUI. There is currently only the right half of the screen displayed which is the two graphs and their dropdowns that they need. Due to there not being anything filled in on the left side for the genetic information (coming soon) everything is spaced closer to center. This will be fixed and respaced once more is added.