Creating Data Processing GUI initial (#1)
* Creating Data Processing GUI initial

This is the first commit of the Data Processing GUI. There is currently only the right half of the screen displayed which is the two graphs and their dropdowns that they need. Due to there not being anything filled in on the left side for the genetic information (coming soon) everything is spaced closer to center. This will be fixed and respaced once more is added.

* New Files and GUI/control Logic

Just saving changes up to this point. New files added for the environment GUI and data processing and environment control. Still several changes needed for adding in functionality and making the GUI look cleaner, add in code comments, and set some boundaries to account for differences in screen sizes.

* Environment Changes

Adding in temporary changes to the environment for both set up and manipulation. Likely to be some change but this is at least the initial idea for the environment for both looks and manipulation

* Secondary idea for Environment

This is a second file that has other ideas for the Environment and how that will be set up and displayed. A lot is commented out just because of it being experimental and playing around with some manipulation of things at the moment.

* New ideas and options

This is saving changes to old files that have been updated and adding in likely the new and final option for creating a grid and buttons in the setup_environmentGUI.py file

* Saving Changes

Saving changes to repo

* Initial GUI creation complete

This is the initial look that the GUI will have. This has the very basics and still has a long ways to go, but it has all pages created, the starting of each of the elements, and all buttons are connected with a function of sorts.

* Small changes and additions

This adds in the ability to fill a square when clicked and updating some docstrings

* Changes for click type

This will allow the user to select a type of   block that is pressed. It hasn't been hooked to the Environment yet, but it does allow the user to select an option still.