Noah/fixing user select click type (#5)
* Creating Data Processing GUI initial

This is the first commit of the Data Processing GUI. There is currently only the right half of the screen displayed which is the two graphs and their dropdowns that they need. Due to there not being anything filled in on the left side for the genetic information (coming soon) everything is spaced closer to center. This will be fixed and respaced once more is added.

* New Files and GUI/control Logic

Just saving changes up to this point. New files added for the environment GUI and data processing and environment control. Still several changes needed for adding in functionality and making the GUI look cleaner, add in code comments, and set some boundaries to account for differences in screen sizes.

* Environment Changes

Adding in temporary changes to the environment for both set up and manipulation. Likely to be some change but this is at least the initial idea for the environment for both looks and manipulation

* Secondary idea for Environment

This is a second file that has other ideas for the Environment and how that will be set up and displayed. A lot is commented out just because of it being experimental and playing around with some manipulation of things at the moment.

* New ideas and options

This is saving changes to old files that have been updated and adding in likely the new and final option for creating a grid and buttons in the setup_environmentGUI.py file

* Saving Changes

Saving changes to repo

* Initial GUI creation complete

This is the initial look that the GUI will have. This has the very basics and still has a long ways to go, but it has all pages created, the starting of each of the elements, and all buttons are connected with a function of sorts.

* Small changes and additions

This adds in the ability to fill a square when clicked and updating some docstrings

* Changes for click type

This will allow the user to select a type of   block that is pressed. It hasn't been hooked to the Environment yet, but it does allow the user to select an option still.

* Minor changes and fixes

Commenting out print statements for the moment and setting the environment click type

* New changes and socket client

There are several todo's that were added just for future notes and future implementation items. Refactoring of the python files structure. Added implementation for button click type still in progress. Added support for socket client to receive information with. Added function to send message data with in the future

* Connection setup

Added functions to establish connection with c++ application, functions to receive and send information and added a tie from the environment to settings being changed in the settings window

* Sending messages

Adding the ability to send messages if some settings are changed. Added some refactoring and addition of button on the settings page to reduce number of messages sent when settings change

* Refactoring and new messages

All messages that will be sent have now been implemented and will be sent to the connected server. Also refactored to have a main function/file as the entry point of the application and have common locations and references to objects