Organism grid (#7)
* Fixing decoding of message that was received to interpret the full message correctly

* Fixed message formats being sent

* latest changes

current changes. Work in progress

* uuid class + organisms actually get serialized

* Refactoring for multithreading

Moved from multiprocessing to multithreading to allow for shared variable access and moved the variables over to a global access file. Also changed the way that dropdowns and spinboxes appear in the settings window

* Temp changes

Work in progress on making the grid able to be changed dynamically

* temp changes and new options

Commented out the procedure for making dynamic environment size for now and then added in support for interpreting and displaying messages. Only tested on one message so far

* new changes

Adding new changes to the way that food types, and organisms are rendered onto the page

* make it build again

* fix whitespace at the beginning of uuid

* fixed receiving of messages

Fixed the way that messages are received so that it actually receives the full message and will also decode certain parts of the grid as signed vs unsigned

* new display changes

Added in the functionality of displaying the organisms with colors.

* initial implementation of organismGrid

* Changed updating of the UI and message types

The way that messages are grouped and sent out was changed to have control be more broad and include speed. Also changed it so that the environment can clear itself just before rendering new organisms.

* Added message box check on exit

Added a popup message box that the user has to click on at exit time in order to confirm that they want to exit the python program. Also added comments to global access to make things a little bit clearer as to what they are

* organism moves around

* ignore venv folder

* remove socket debug print statements

* ignore pycache

* actually ignore .pyc

---------

Co-authored-by: Noah <hudkinsn@oregonstate.edu>